defautl html ok

+++ src/KPEEPROMClass.cpp
@@ -14,31 +14,31 @@ KPEEPROMClass::KPEEPROMClass()
 {
 }
 
-int8_t KPEEPROMClass::getEEPROMInt8(byte offset)
+int8_t KPEEPROMClass::getEEPROMInt8(int offset)
 {
   // Le dado Int8 da KPEEPROMClass::EEPROMClass::(–128 a 127)
   return int8(KPEEPROMClass::EEPROMClass::read(offset));
 }
 
-uint8_t KPEEPROMClass::getEEPROMUInt8(byte offset)
+uint8_t KPEEPROMClass::getEEPROMUInt8(int offset)
 {
   // Le dado UInt8 da KPEEPROMClass::EEPROMClass::(0 a 255)
   return KPEEPROMClass::EEPROMClass::read(offset);
 }
 
-int16_t KPEEPROMClass::getEEPROMInt16(byte offset)
+int16_t KPEEPROMClass::getEEPROMInt16(int offset)
 {
   // Le dado Int16 da KPEEPROMClass::EEPROMClass::(–32768 a 32767)
   return short(word(KPEEPROMClass::EEPROMClass::read(offset), KPEEPROMClass::EEPROMClass::read(offset + 1)));
 }
 
-uint16_t KPEEPROMClass::getEEPROMUInt16(byte offset)
+uint16_t KPEEPROMClass::getEEPROMUInt16(int offset)
 {
   // Le dado UInt16 da KPEEPROMClass::EEPROMClass::(0 a 65535)
   return word(KPEEPROMClass::EEPROMClass::read(offset), KPEEPROMClass::EEPROMClass::read(offset + 1));
 }
 
-int32_t KPEEPROMClass::getEEPROMInt32(byte offset)
+int32_t KPEEPROMClass::getEEPROMInt32(int offset)
 {
   // Le dado Int32 da KPEEPROMClass::EEPROMClass::(–2.147.483.648 a 2.147.483.647)
   int32 l;
@@ -52,7 +52,7 @@ int32_t KPEEPROMClass::getEEPROMInt32(byte offset)
   return l;
 }
 
-uint32_t KPEEPROMClass::getEEPROMUInt32(byte offset)
+uint32_t KPEEPROMClass::getEEPROMUInt32(int offset)
 {
   // Le dado UInt32 da KPEEPROMClass::EEPROMClass::(0 a 4.294.967.295)
   uint32 l;
@@ -66,7 +66,7 @@ uint32_t KPEEPROMClass::getEEPROMUInt32(byte offset)
   return l;
 }
 
-String KPEEPROMClass::getEEPROMString(byte offset, word memSize)
+String KPEEPROMClass::getEEPROMString(int offset, word memSize)
 {
   // Le String da EEPROM
   String s = "";
@@ -82,33 +82,33 @@ String KPEEPROMClass::getEEPROMString(byte offset, word memSize)
   return s;
 }
 
-void KPEEPROMClass::setEEPROMInt8(byte offset, int8_t v)
+void KPEEPROMClass::setEEPROMInt8(int offset, int8_t v)
 {
   // Grava dado Int8 na KPEEPROMClass::EEPROMClass::(–128 a 127)
   KPEEPROMClass::EEPROMClass::write(offset, v);
 }
 
-void KPEEPROMClass::setEEPROMInt8(byte offset, uint8_t v)
+void KPEEPROMClass::setEEPROMInt8(int offset, uint8_t v)
 {
   // Grava dado UInt8 na KPEEPROMClass::EEPROMClass::(0 a 255)
   KPEEPROMClass::EEPROMClass::write(offset, v);
 }
 
-void KPEEPROMClass::setEEPROMInt16(byte offset, int16_t v)
+void KPEEPROMClass::setEEPROMInt16(int offset, int16_t v)
 {
   // Grava dado Int16 na KPEEPROMClass::EEPROMClass::(–32768 a 32767)
   KPEEPROMClass::EEPROMClass::write(offset, highByte(v));
   KPEEPROMClass::EEPROMClass::write(offset + 1, lowByte(v));
 }
 
-void KPEEPROMClass::setEEPROMInt16(byte offset, uint16_t v)
+void KPEEPROMClass::setEEPROMInt16(int offset, uint16_t v)
 {
   // Grava dado UInt16 na KPEEPROMClass::EEPROMClass::(0 a 65535)
   KPEEPROMClass::EEPROMClass::write(offset, highByte(v));
   KPEEPROMClass::EEPROMClass::write(offset + 1, lowByte(v));
 }
 
-void KPEEPROMClass::setEEPROMInt32(byte offset, int32_t v)
+void KPEEPROMClass::setEEPROMInt32(int offset, int32_t v)
 {
   // Grava dado Int32 na KPEEPROMClass::EEPROMClass::(–2147483648 a 2147483647)
   KPEEPROMClass::EEPROMClass::write(offset + 3, byte(v));
@@ -120,7 +120,7 @@ void KPEEPROMClass::setEEPROMInt32(byte offset, int32_t v)
   KPEEPROMClass::EEPROMClass::write(offset, byte(v));
 }
 
-void KPEEPROMClass::setEEPROMInt32(byte offset, uint32_t v)
+void KPEEPROMClass::setEEPROMInt32(int offset, uint32_t v)
 {
   // Grava dado UInt32 na KPEEPROMClass::EEPROMClass::(0 a 4294967295)
   KPEEPROMClass::EEPROMClass::write(offset + 3, byte(v));
@@ -132,7 +132,7 @@ void KPEEPROMClass::setEEPROMInt32(byte offset, uint32_t v)
   KPEEPROMClass::EEPROMClass::write(offset, byte(v));
 }
 
-void KPEEPROMClass::setEEPROMString(byte offset, String s)
+void KPEEPROMClass::setEEPROMString(int offset, String s)
 {
   // Grava String na EEPROM
   for (byte b = 0; b <= s.length(); b++)
